Common methods each and self.subclass for collections and subcollections moved to the common collection mixin
