+++ lib/manageiq/api/client/subcollection.rb
@@ -22,20 +22,6 @@ module ManageIQ
           fetch_actions(result_hash)
         end
 
-        def each(&block)
-          all.each(&block)
-        end
-
-        def self.subclass(name)
-          name = name.camelize
-
-          if const_defined?(name, false)
-            const_get(name, false)
-          else
-            const_set(name, Class.new(self))
-          end
-        end
-
         def get(options = {})
           options[:expand] = (String(options[:expand]).split(",") | %w(resources)).join(",")
           options[:filter] = Array(options[:filter]) if options[:filter].is_a?(String)
